moved dd fetcher inside usageaccountant to have one top-level package
